requirements.txt
streamlit
langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-text-splitters==0.2.2
langchain-google-genai
langchain-huggingface
sentence-transformers
google-generativeai
chromadb>=0.4.24
pydantic>=2.5.0
pymupdf
numpy
python-dotenv